pyproject.toml
[project]
name = "squeeze-tw-futures-realtime"
version = "0.1.0"
description = "Real-time Squeeze Momentum strategy for Taiwan Index Futures"
requires-python = ">=3.12"
authors = [
    {name = "mylin102"}
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pandas-ta>=0.3.14b0",
    "yfinance>=0.2.36",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "shioaji[speed]>=1.3.2",
    "tabulate>=0.10.0",
    "pyyaml>=6.0.3",
    "jinja2>=3.1.6",
]

[project.scripts]
squeeze-futures = "squeeze_futures.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
